+++ spec/factories/event.rb
@@ -14,5 +14,18 @@ FactoryBot.define do
         create_list(:attendance, evaluator.attendances_count, event: event)
       end
     end
+
+    factory :event_with_preregistrations do
+      preregistration { true }
+
+      transient do
+        preregistrations_count { 5 }
+      end
+
+      after(:create) do |event, evaluator|
+        create_list(:preregistration, evaluator.preregistrations_count,
+                    event: event)
+      end
+    end
   end
 end